make forge and neoforge concurrent
