+++ meta/common/__init__.py
@@ -95,3 +95,11 @@ def remove_files(file_paths):
                 os.remove(file_path)
         except Exception as e:
             print(e)
+
+
+def filehash(filename, hashtype, blocksize=65536):
+    hashtype = hashtype()
+    with open(filename, "rb") as f:
+        for block in iter(lambda: f.read(blocksize), b""):
+            hashtype.update(block)
+    return hashtype.hexdigest()
